// client/src/app/components/types/button.types.ts
export type variant = 'filled' | 'outlined' | 'text';


export type size = 'sm' | 'md' | 'lg';

export type colors =
  | 'ds-primary'
  | 'ds-secondary'
  | 'primary'
  | 'secondary'
  | 'slate'
  | 'gray'
  | 'zinc'
  | 'neutral'
  | 'red'
  | 'yellow'
  | 'green';

export type colorVaraints =
  | 0
  | 50
  | 100
  | 200
  | 300
  | 400
  | 500
  | 600
  | 700
  | 800
  | 900
  | 950;

  export type roundedSize = 'sm' | 'md' | 'lg' | 'full';

  export type borderStyle = 'dashed' | 'dotted' | 'double' | 'solid' | 'none';